Use 5d2214e324 of bbroeksema/parallel-coordinates.
This has a good chance of becoming master in:
syntagmatic/parallel-coordinates

This has a better implementation of the color index.

It also contains Kai's temporary revert of my chance which returned false
for the selection when all brushes where cleared. Therefore, we need to
adapt the data indicator contoller a bit.