Divide the window into two rows
A bootstrap row takes full with by default, so the only thing we need to
adjust is the height. We create one top row for the parallel coordinates
and a bottom row for the controls and the scatter plot. This is a first
stab at properly setting width and height of divs, without hacks.