Add a first version of the (new) DimRedPlot
Reasoning: The primary (and probably only) goal of this plot will be to add
some functionality based on *column* projections. Additionally, plotting
row points is redundant as they can be added to the parallel coordinates.
Finally, I want a component specific for (P)CA as it is likely that I'll
soon start a collaboration with my former supervisor.