Extend the functionality of the DimRedPlot.
We now show a barplot on the left, which visualizes all the principal axes
and the amount of explained variance.

Using the scroll wheel on this barplot will trigger an "rotate" event,
which has one argument that takes one of ["up", "down"]. This can be used
to change the "active" pair of principal axes (i.e. the ones which are
currently visualized in the plot).