Be more specific about what to select.
The rectangles representing values are classed as 'point'. Reflect this in
the selection to make clear that render.points is *only* about changing
those (and not potentially other rectangles, which are added for other
purposes).

I had this problem when I had an overlay rectangle to capture scroll wheel
events, which was styled differently and thus would give 'funny' effects on
updates. Though, this code is no longer there, I figured it still makes
sense to be specific, to avoid this problem in the future.