Emit an event when a point is clicked.
Somehow, we need to inform the backend how the new variables, based on TNF
clusters need to be formed. To this extend we emit an event when a point is
clicked. The event will have one argument, an array that contains the
clustered tertra nucleotides. In case of no clustering, it will be an array
with one element.