Class the point rectangles as soon as they are appended.
The class of the point rectangles is not something which changes based on
data (at least not the '.point' part), so we directly add it to the
elements.

This allows for making selections like: gPoints.selectAll("rect.point"),
even before the transition is finished, and it will give all the points.