Add a backend function for summarizing frequency data.
The basic idea is as follows: We use correspondence analysis to project the
nucleotide frequencies. Next, we cluster frequencies that are closely
related on a given pair of principal axes. Finally, we use the weighted sum
of the nucleotides as a summary of a cluster of related nucleotides. The
weights are determined by the contributions of a given nucleotide to the
principal axes that where used for the clustering.