Fix dynamic execution of CA for subsets of contigs.
The initial approach was to store the results of CA and next call
dimred.ca.plotdata from the frontend. However, the results of CA gets big
when it's performed on all contigs, as it will store (among other things)
an N*M matrix, where N is the number of contigs, and M is the number of
Tetra Nucleotides used for the CA. In our current approach we're not using
this matrix at all and I got some opencpu related issues when it tried to
store it in the session file.

In the new approach I call the plotdata function in dimred.ca which
significantly reduces the amount of data that needs to be stored in the
session.