Fix labeling on CA from backend.
Apparently the ca.json file had a slightly different field for the label
than what was produced by the last version. I now added an explicit label
property to the projection data and the DimRedPlot now assumes a label
property in stead of _row.

This could be made more flexible by telling the DimRedPlot which property
must be used for labels in stead of hard-coding it in the plot.