Initial code for an sqlite backend.
The current plan is to use an sqlite database to load the data from. This
saves reading and writing of large rda files by opencpu for most cases. It
also enables adding dynamic columns because this now should simply result
in the addition of update of a column to a database file.

This way we hope to be able to get rid of the complex logic in DataSet.get
in the front-end, and to enable the inclusion of dynamic columns in
clustering and coloring.