Disable including clustering results for pca and ca.
This was a first stab at fixing memory issues, encountered with a
new data set for Magda. Although, in the end this was not actually
the problem, I do think that this is something that should happen
in the frontend.

As a consequence, I also disable the clustering tab for now in the
front end.