First version of a 2D map and its controller.
Some data sets have explicit 2D spatial variables (e.g. longtitude and
latitude). We actually want to display those in a map, instead of as axes
in the parcoords in order to support spatial interpretation of selections
in the parcoords. This commit brings in an intial implementation of such a
map and the required controller. It is not wired with the rest of the
application yet.