First step in moving filtering to the front end.
To make the filtering experience snappier we have to move it to the front-
end. To this end we remove the backend function data.filter. In the front
end we adjusted the filter function to take over this functionality. For
now this doesn't speed anything up, as after an Dataset::filtered signal,
still a request is done to retrieve the filtered data.

The next step is to actually keep the data table in Dataset and perform the
filtering directly on the table.