First stab at moving coloring to the frontend and fix frontend bug.
In this commit we move the available color configurations to the front end
and remove them from the backend. The actual coloring is still performed by
the backend though.

This commit also fixes the bug that the schemes where not properly updated
when the scheme for one method was also available for another method. That
is, the method "value" is available both for numeric and factor types. The
schemes for both are different though.