Fix the decile coloring.
What we implemented was a wrong understunding of me of decile coloring. I
actually splitted the value range into ten equal bands, while it was
supposed to split the items into ten (almost) equal bins.

So, now we first sort the data items on color variable. Next we assign the
color to the current item based on the current bin. Bin number is increased
when the required bin size is reached *and* when the current value is not
the same as the previous value.