Fix dim red bug when #row < #vars
When applying dimensionality reduction to a dataset with a smaller number
of rows than variables, one gets only as many column (or variable)
projections as rows. This resulted in the (P|M)?CA plot not showing up when
the user had filtered down to a small number of rows and tried to do dim
red on a larger number of variables.