Changed the brush broadcast system to be more flexible
Now when an element wants to brush it has to call brush on the DataSet
service while passing along an identifier identifying the caller.
DataSet will broadcast that a brush happened and receivers of the brush
broadcast can choose to ignore the brush event if they were the ones
making it.