Updated ui-bootstrap to version 12.1
This was necessary for a feature in the next commit. I also fixed a bug
here, you are not supposed to include bootstrap.js yourself. Instead,
ui-bootstrap should do this for you. This caused malfunction with
tabbing so I removed the referenced bootstrap.js and rewrote the tabbing
to be ui-bootstrap compatible.