Added PCA support. Closes #5.
PCA support is added. This means that support is now available for
showing both PCA and CA next to each other. When filtering is applied,
both PCA and CA are updated, but on the variables that were used to
initialise them. To achieve this, the DimRedPlot service is now
responsible for performing PCA and CA, such that it also knows which
analyses have been performed. The effect of this change is also that
some of the logic from plotctrl has been moved into the DimRedPlot
service, which is good since controllers shouldn't have too much logic
anyways.